OpenAIAPIの呼び出しをResponseコントローラーからResponseモデル移行しました
